requirements.txt
# Core dependencies
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.1.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# API and utilities
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional but recommended
tiktoken>=0.5.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.0.0

# Web interface
streamlit>=1.28.0
